kuwala/core/database/importer/requirements.txt
psycopg2-binary==2.9.1
py4j==0.10.9.2
pyspark==3.2.0
